IdP tests are now more DRY
