+++ src/test/java/org/folio/rest/impl/IdpCallbackTest.java
@@ -16,20 +16,20 @@ import org.testcontainers.containers.output.Slf4jLogConsumer;
  * Test against a real IDP: https://simplesamlphp.org/ running in a Docker container.
  */
 @RunWith(VertxUnitRunner.class)
-public class IdpTest {
+public class IdpCallbackTest {
   private static final org.slf4j.Logger logger = LoggerFactory.getLogger(IdpTest.class);
   private static final boolean DEBUG = false;
   private static final String TENANT = SamlTestHelper.TENANT;
   private static final int MODULE_PORT = SamlTestHelper.MODULE_PORT;
   private static final String OKAPI_URL = SamlTestHelper.OKAPI_URL;
-  private static final String CALLBACK_WITH_EXPIRY = "callback-with-expiry";
+  private static final String CALLBACK = "callback";
   private static Vertx VERTX;
 
   @ClassRule
-  public static final SimpleSamlPhpContainer<?> IDP = new SimpleSamlPhpContainer<>(OKAPI_URL, CALLBACK_WITH_EXPIRY);
+  public static final SimpleSamlPhpContainer<?> IDP = new SimpleSamlPhpContainer<>(OKAPI_URL, CALLBACK);
 
   @BeforeClass
-  public static void setupOnce(TestContext context) {
+  public static void setupOnce(TestContext context)  {
     RestAssured.port = MODULE_PORT;
     RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
     VERTX = Vertx.vertx();
@@ -37,7 +37,6 @@ public class IdpTest {
     if (DEBUG) {
       IDP.followOutput(new Slf4jLogConsumer(logger).withSeparateOutputStreams());
     }
-
     IDP.init();
 
     SamlTestHelper.deployVerticle(VERTX, context);
@@ -46,7 +45,7 @@ public class IdpTest {
   @AfterClass
   public static void tearDownOnce(TestContext context) {
     VERTX.close()
-    .onComplete(context.asyncAssertSuccess());
+      .onComplete(context.asyncAssertSuccess());
   }
 
   @After
@@ -55,22 +54,22 @@ public class IdpTest {
   }
 
   @Test
-  public void post() {
+  public void postCallback() {
     IDP.setPostBinding();
-    SamlTestHelper.setOkapi("mock_idptest_post.json", IDP);
+    SamlTestHelper.setOkapi("mock_idptest_post_secure_tokens.json", IDP);
 
     for (int i = 0; i < 2; i++) {
-      SamlTestHelper.testPost(CALLBACK_WITH_EXPIRY);
+      SamlTestHelper.testPost(CALLBACK);
     }
   }
 
   @Test
-  public void redirect() {
+  public void redirectCallback() {
     IDP.setRedirectBinding();
-    SamlTestHelper.setOkapi("mock_idptest_redirect.json", IDP);
+    SamlTestHelper.setOkapi("mock_idptest_redirect_secure_tokens.json", IDP);
 
     for (int i = 0; i < 2; i++) {
-      SamlTestHelper.testRedirect(CALLBACK_WITH_EXPIRY);
+      SamlTestHelper.testRedirect(CALLBACK);
     }
   }
 }
