+++ src/test/java/org/folio/rest/impl/IdpLegacyTest.java
@@ -16,12 +16,11 @@ import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
-import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.nio.file.Path;
 import java.util.regex.Pattern;
 import org.folio.config.SamlConfigHolder;
+import org.folio.testutil.SimpleSamlPhpContainer;
 import org.folio.rest.RestVerticle;
 import org.folio.util.MockJson;
 import org.folio.util.StringUtil;
@@ -32,9 +31,7 @@ import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.slf4j.LoggerFactory;
-import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.output.Slf4jLogConsumer;
-import org.testcontainers.images.builder.ImageFromDockerfile;
 
 /**
  * Test against a real IDP: https://simplesamlphp.org/ running in a Docker container.
@@ -43,35 +40,26 @@ import org.testcontainers.images.builder.ImageFromDockerfile;
 public class IdpLegacyTest {
   private static final org.slf4j.Logger logger = LoggerFactory.getLogger(IdpLegacyTest.class);
   private static final boolean DEBUG = false;
-  private static final ImageFromDockerfile simplesamlphp =
-      new ImageFromDockerfile().withFileFromPath(".", Path.of("src/test/resources/simplesamlphp/"));
-
   private static final String TENANT = "diku";
   private static final Header TENANT_HEADER = new Header("X-Okapi-Tenant", TENANT);
   private static final Header TOKEN_HEADER = new Header("X-Okapi-Token", "mytoken");
   private static final Header JSON_CONTENT_TYPE_HEADER = new Header("Content-Type", "application/json");
   private static final String STRIPES_URL = "http://localhost:3000";
-
   private static final int MODULE_PORT = 9231;
   private static final String MODULE_URL = "http://localhost:" + MODULE_PORT;
   private static final int OKAPI_PORT = 9230;
   private static final String OKAPI_URL = "http://localhost:" + OKAPI_PORT;
-  private static int IDP_PORT;
-  private static String IDP_BASE_URL;
   private static final Header OKAPI_URL_HEADER = new Header("X-Okapi-Url", OKAPI_URL);
   private static MockJson OKAPI;
 
   private static Vertx VERTX;
 
   @ClassRule
-  public static final GenericContainer<?> IDP = new GenericContainer<>(simplesamlphp)
-      .withExposedPorts(8080)
-      .withEnv("SIMPLESAMLPHP_SP_ENTITY_ID", OKAPI_URL + "/_/invoke/tenant/diku/saml/callback")
-      .withEnv("SIMPLESAMLPHP_SP_ASSERTION_CONSUMER_SERVICE",
-               OKAPI_URL + "/_/invoke/tenant/diku/saml/callback");
+  public static final SimpleSamlPhpContainer<?> IDP =
+      new SimpleSamlPhpContainer<>(OKAPI_URL, "callback");
 
   @BeforeClass
-  public static void setupOnce(TestContext context) throws Exception {
+  public static void setupOnce(TestContext context) {
     RestAssured.port = MODULE_PORT;
     RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
     VERTX = Vertx.vertx();
@@ -79,13 +67,8 @@ public class IdpLegacyTest {
     if (DEBUG) {
       IDP.followOutput(new Slf4jLogConsumer(logger).withSeparateOutputStreams());
     }
-    IDP_PORT = IDP.getFirstMappedPort();
-    IDP_BASE_URL = "http://" + IDP.getHost() + ":" + IDP_PORT + "/simplesaml/";
-    String baseurlpath = IDP_BASE_URL.replace("/", "\\/");
-    exec("sed", "-i", "s/'baseurlpath' =>.*/'baseurlpath' => '" + baseurlpath + "',/",
-        "/var/www/simplesamlphp/config/config.php");
-    exec("sed", "-i", "s/'auth' =>.*/'auth' => 'example-static',/",
-        "/var/www/simplesamlphp/metadata/saml20-idp-hosted.php");
+
+    IDP.init();
 
     DeploymentOptions moduleOptions = new DeploymentOptions()
         .setConfig(new JsonObject().put("http.port", MODULE_PORT)
@@ -113,7 +96,7 @@ public class IdpLegacyTest {
 
   @Test
   public void post() {
-    setIdpBinding("POST");
+    IDP.setPostBinding();;
     setOkapi("mock_idptest_post_legacy.json");
 
     for (int i = 0; i < 2; i++) {
@@ -169,7 +152,7 @@ public class IdpLegacyTest {
 
   @Test
   public void redirect() {
-    setIdpBinding("Redirect");
+    IDP.setRedirectBinding();
     setOkapi("mock_idptest_redirect_legacy.json");
 
     for (int i = 0; i < 2; i++) {
@@ -235,28 +218,8 @@ public class IdpLegacyTest {
       header("Location", startsWith("http://localhost:3000/sso-landing?ssoToken=saml-token"));
   }
 
-  private void setIdpBinding(String binding) {
-    // append entry at end, last entry wins
-    exec("sed", "-i",
-        "s/];/'SingleSignOnServiceBinding' => 'urn:oasis:names:tc:SAML:2.0:bindings:HTTP-" + binding + "',\\n];/",
-        "/var/www/simplesamlphp/metadata/saml20-idp-hosted.php");
-  }
-
-  private static void exec(String... command) {
-    try {
-      var result = IDP.execInContainer(command);
-      if (result.getExitCode() > 0) {
-        System.out.println(result.getStdout());
-        System.err.println(result.getStderr());
-          throw new RuntimeException("failure in IDP.execInContainer");
-      }
-    } catch (UnsupportedOperationException | IOException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
   private void setOkapi(String resource) {
-    OKAPI.setMockContent(resource, s -> s.replace("http://localhost:8888/simplesaml/", IDP_BASE_URL));
+    OKAPI.setMockContent(resource, s -> s.replace("http://localhost:8888/simplesaml/", IDP.getBaseUrl()));
   }
 
   private String jsonEncode(String key, String value) {
